# PERF: improve perf. of Categorical.searchsorted